# kenyan-real-estate-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Amenities:
    additionalProperties: true
    type: object
...
  models.CreatePropertyRequest:
    properties:
      amenities:
        $ref: '#/definitions/models.Amenities'
      availability_date:
        type: string
      bathrooms:
        minimum: 0
        type: integer
      bedrooms:
        minimum: 0
        type: integer
      county_id:
        type: integer
      deposit_amount:
        minimum: 0
        type: number
      description:
        type: string
      is_furnished:
        type: boolean
      latitude:
        type: number
      location_details:
        type: string
      longitude:
        type: number
      parking_spaces:
        minimum: 0
        type: integer
      property_type:
        $ref: '#/definitions/models.PropertyType'
      rent_amount:
        minimum: 0
        type: number
      square_meters:
        minimum: 0
        type: number
      sub_county_id:
        type: integer
      title:
        type: string
      utilities_included:
        $ref: '#/definitions/models.UtilitiesIncluded'
    required:
    - county_id
    - property_type
    - rent_amount
    - title
    type: object
...
  models.Property:
    properties:
      amenities:
        $ref: '#/definitions/models.Amenities'
      availability_date:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      county:
        allOf:
        - $ref: '#/definitions/models.County'
        description: Relationships
      county_id:
        type: integer
      created_at:
        type: string
      deposit_amount:
        type: number
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.PropertyImage'
        type: array
      is_available:
        type: boolean
      is_furnished:
        type: boolean
      landlord:
        $ref: '#/definitions/models.User'
      landlord_id:
        type: string
      latitude:
        type: number
      location_details:
        type: string
      longitude:
        type: number
      parking_spaces:
        type: integer
      property_type:
        $ref: '#/definitions/models.PropertyType'
      rent_amount:
        type: number
      square_meters:
        type: number
      sub_county:
        $ref: '#/definitions/models.SubCounty'
      sub_county_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      utilities_included:
        $ref: '#/definitions/models.UtilitiesIncluded'
    type: object
...
  models.UtilitiesIncluded:
    additionalProperties: true
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for managing real estate properties in Kenya
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kenyan Real Estate API
  version: "1.0"
paths:
...
  /my-properties:
    get:
      consumes:
      - application/json
      description: Get all properties owned by the authenticated landlord
      parameters:
      - default: 20
        description: Number of results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of landlord's properties
          schema:
            properties:
              properties:
                items:
                  $ref: '#/definitions/models.Property'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get my properties
      tags:
      - Properties
...
  /properties:
    get:
      consumes:
      - application/json
      description: Get a list of available properties with optional filtering and
        pagination
      parameters:
      - description: Filter by county ID
        in: query
        name: county_id
        type: integer
      - description: Filter by sub-county ID
        in: query
        name: sub_county_id
        type: integer
      - description: Filter by property type
        enum:
        - apartment
        - house
        - bedsitter
        - studio
        - maisonette
        - bungalow
        - villa
        - commercial
        in: query
        name: property_type
        type: string
      - description: Minimum rent amount
        in: query
        name: min_rent
        type: number
      - description: Maximum rent amount
        in: query
        name: max_rent
        type: number
      - description: Minimum number of bedrooms
        in: query
        name: min_bedrooms
        type: integer
      - description: Maximum number of bedrooms
        in: query
        name: max_bedrooms
        type: integer
      - description: Minimum number of bathrooms
        in: query
        name: min_bathrooms
        type: integer
      - description: Filter by furnished status
        in: query
        name: is_furnished
        type: boolean
      - description: Filter by parking availability
        in: query
        name: has_parking
        type: boolean
      - default: 20
        description: Number of results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of properties
          schema:
            properties:
              limit:
                type: integer
              offset:
                type: integer
              properties:
                items:
                  $ref: '#/definitions/models.Property'
                type: array
              total:
                type: integer
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get public property listings
      tags:
      - Properties
    post:
      consumes:
      - application/json
      description: Create a new property listing (landlord only)
      parameters:
      - description: Property data
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/models.CreatePropertyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Property created successfully
          schema:
            properties:
              message:
                type: string
              property:
                $ref: '#/definitions/models.Property'
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new property
      tags:
      - Properties
  /properties/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a property (landlord only)
      parameters:
      - description: Property ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property deleted successfully
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Invalid property ID
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: You can only delete your own properties
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Property not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Delete property
      tags:
      - Properties
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific property including images
      parameters:
      - description: Property ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property details
          schema:
            properties:
              property:
                $ref: '#/definitions/models.Property'
            type: object
        "400":
          description: Invalid property ID
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Property not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a property by ID
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Update property information (landlord only)
      parameters:
      - description: Property ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Property update data
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Property updated successfully
          schema:
            properties:
              message:
                type: string
              property:
                $ref: '#/definitions/models.Property'
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: You can only update your own properties
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Property not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Update property
      tags:
      - Properties
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"